+++ decoder_modules/radio/src/demodulators/am.h
@@ -25,13 +25,22 @@ namespace demod {
             if (config->conf[name][getName()].contains("agcDecay")) {
                 agcDecay = config->conf[name][getName()]["agcDecay"];
             }
-            if (config->conf[name][getName()].contains("carrierAgc")) {
-                carrierAgc = config->conf[name][getName()]["carrierAgc"];
+            if (config->conf[name][getName()].contains("agcMode")) {
+                agcMode = std::clamp<int>(config->conf[name][getName()]["agcMode"], 0, 2);
+            }
+            else if (config->conf[name][getName()].contains("carrierAgc")) {
+                // Legacy config format with a carrier AGC on/off flag
+                bool carrierAgc = config->conf[name][getName()]["carrierAgc"];
+                agcMode = carrierAgc ? dsp::demod::AM<dsp::stereo_t>::AGCMode::CARRIER : dsp::demod::AM<dsp::stereo_t>::AGCMode::AUDIO;
+            }
+            if (config->conf[name][getName()].contains("agcGain")) {
+                agcGain = config->conf[name][getName()]["agcGain"];
             }
             config->release();
 
             // Define structure
-            demod.init(input, carrierAgc ? dsp::demod::AM<dsp::stereo_t>::AGCMode::CARRIER : dsp::demod::AM<dsp::stereo_t>::AGCMode::AUDIO, bandwidth, agcAttack / getIFSampleRate(), agcDecay / getIFSampleRate(), 100.0 / getIFSampleRate(), getIFSampleRate());
+            demod.init(input, (dsp::demod::AM<dsp::stereo_t>::AGCMode)agcMode, bandwidth, agcAttack / getIFSampleRate(), agcDecay / getIFSampleRate(), 100.0 / getIFSampleRate(), getIFSampleRate());
+            demod.setAGCGain(powf(10.0f, agcGain / 20.0f));
         }
 
         void start() { demod.start(); }
@@ -40,6 +49,34 @@ namespace demod {
 
         void showMenu() {
             float menuWidth = ImGui::GetContentRegionAvail().x;
+            ImGui::LeftLabel("AGC Mode");
+            ImGui::SetNextItemWidth(menuWidth - ImGui::GetCursorPosX());
+            if (ImGui::Combo(("##_radio_am_agc_mode_" + name).c_str(), &agcMode, "Off\0Carrier\0Audio\0")) {
+                demod.setAGCMode((dsp::demod::AM<dsp::stereo_t>::AGCMode)agcMode);
+                agcGain = std::clamp<float>(20.0f * log10f(demod.getAGCGain()), -10.0f, 90.0f);
+                _config->acquire();
+                _config->conf[name][getName()]["agcMode"] = agcMode;
+                if (agcMode == dsp::demod::AM<dsp::stereo_t>::AGCMode::OFF) {
+                    // Keep the last AGC gain as the manual gain
+                    _config->conf[name][getName()]["agcGain"] = agcGain;
+                }
+                _config->release(true);
+            }
+            bool agcEnabled = (agcMode != dsp::demod::AM<dsp::stereo_t>::AGCMode::OFF);
+            if (agcEnabled) {
+                agcGain = std::clamp<float>(20.0f * log10f(demod.getAGCGain()), -10.0f, 90.0f);
+                ImGui::BeginDisabled();
+            }
+            ImGui::LeftLabel("Gain");
+            ImGui::SetNextItemWidth(menuWidth - ImGui::GetCursorPosX());
+            if (ImGui::SliderFloat(("##_radio_am_gain_" + name).c_str(), &agcGain, -10.0f, 90.0f, "%.0f dB")) {
+                demod.setAGCGain(powf(10.0f, agcGain / 20.0f));
+                _config->acquire();
+                _config->conf[name][getName()]["agcGain"] = agcGain;
+                _config->release(true);
+            }
+            if (agcEnabled) { ImGui::EndDisabled(); }
+            else { ImGui::BeginDisabled(); }
             ImGui::LeftLabel("AGC Attack");
             ImGui::SetNextItemWidth(menuWidth - ImGui::GetCursorPosX());
             if (ImGui::SliderFloat(("##_radio_am_agc_attack_" + name).c_str(), &agcAttack, 1.0f, 200.0f)) {
@@ -56,12 +93,7 @@ namespace demod {
                 _config->conf[name][getName()]["agcDecay"] = agcDecay;
                 _config->release(true);
             }
-            if (ImGui::Checkbox(("Carrier AGC##_radio_am_carrier_agc_" + name).c_str(), &carrierAgc)) {
-                demod.setAGCMode(carrierAgc ? dsp::demod::AM<dsp::stereo_t>::AGCMode::CARRIER : dsp::demod::AM<dsp::stereo_t>::AGCMode::AUDIO);
-                _config->acquire();
-                _config->conf[name][getName()]["carrierAgc"] = carrierAgc;
-                _config->release(true);
-            }
+            if (!agcEnabled) { ImGui::EndDisabled(); }
         }
 
         void setBandwidth(double bandwidth) { demod.setBandwidth(bandwidth); }
@@ -93,9 +125,10 @@ namespace demod {
 
         ConfigManager* _config = NULL;
 
+        int agcMode = dsp::demod::AM<dsp::stereo_t>::AGCMode::AUDIO;
+        float agcGain = 0.0f;
         float agcAttack = 50.0f;
         float agcDecay = 5.0f;
-        bool carrierAgc = false;
 
         std::string name;
     };