+++ decoder_modules/radio/src/demodulators/cwr.h
@@ -21,6 +21,12 @@ namespace demod {
             this->afbwChangeHandler = afbwChangeHandler;
 
             config->acquire();
+            if (config->conf[name][getName()].contains("agcEnabled")) {
+                agcEnabled = config->conf[name][getName()]["agcEnabled"];
+            }
+            if (config->conf[name][getName()].contains("agcGain")) {
+                agcGain = config->conf[name][getName()]["agcGain"];
+            }
             if (config->conf[name][getName()].contains("agcAttack")) {
                 agcAttack = config->conf[name][getName()]["agcAttack"];
             }
@@ -33,6 +39,8 @@ namespace demod {
             config->release();
 
             demod.init(input, -tone, agcAttack / getIFSampleRate(), agcDecay / getIFSampleRate(), getIFSampleRate());
+            demod.setAGCEnabled(agcEnabled);
+            demod.setAGCGain(powf(10.0f, agcGain / 20.0f));
         }
 
         void start() { demod.start(); }
@@ -41,6 +49,30 @@ namespace demod {
 
         void showMenu() {
             float menuWidth = ImGui::GetContentRegionAvail().x;
+            if (ImGui::Checkbox(("AGC##_radio_cwr_agc_ena_" + name).c_str(), &agcEnabled)) {
+                demod.setAGCEnabled(agcEnabled);
+                _config->acquire();
+                _config->conf[name][getName()]["agcEnabled"] = agcEnabled;
+                if (!agcEnabled) {
+                    // Keep the last AGC gain as the manual gain
+                    _config->conf[name][getName()]["agcGain"] = agcGain;
+                }
+                _config->release(true);
+            }
+            if (agcEnabled) {
+                agcGain = std::clamp<float>(20.0f * log10f(demod.getAGCGain()), -10.0f, 90.0f);
+                ImGui::BeginDisabled();
+            }
+            ImGui::SameLine();
+            ImGui::SetNextItemWidth(menuWidth - ImGui::GetCursorPosX());
+            if (ImGui::SliderFloat(("##_radio_cwr_gain_" + name).c_str(), &agcGain, -10.0f, 90.0f, "%.0f dB")) {
+                demod.setAGCGain(powf(10.0f, agcGain / 20.0f));
+                _config->acquire();
+                _config->conf[name][getName()]["agcGain"] = agcGain;
+                _config->release(true);
+            }
+            if (agcEnabled) { ImGui::EndDisabled(); }
+            else { ImGui::BeginDisabled(); }
             ImGui::LeftLabel("AGC Attack");
             ImGui::SetNextItemWidth(menuWidth - ImGui::GetCursorPosX());
             if (ImGui::SliderFloat(("##_radio_cwr_agc_attack_" + name).c_str(), &agcAttack, 1.0f, 200.0f)) {
@@ -57,6 +89,7 @@ namespace demod {
                 _config->conf[name][getName()]["agcDecay"] = agcDecay;
                 _config->release(true);
             }
+            if (!agcEnabled) { ImGui::EndDisabled(); }
             ImGui::LeftLabel("Tone Frequency");
             ImGui::FillWidth();
             if (ImGui::InputInt(("Stereo##_radio_cwr_tone_" + name).c_str(), &tone, 10, 100)) {
@@ -96,6 +129,8 @@ namespace demod {
 
         std::string name;
 
+        bool agcEnabled = true;
+        float agcGain = 0.0f;
         float agcAttack = 100.0f;
         float agcDecay = 5.0f;
         int tone = 800;