+++ decoder_modules/radio/src/demodulators/lsb.h
@@ -21,6 +21,12 @@ namespace demod {
 
             // Load config
             config->acquire();
+            if (config->conf[name][getName()].contains("agcEnabled")) {
+                agcEnabled = config->conf[name][getName()]["agcEnabled"];
+            }
+            if (config->conf[name][getName()].contains("agcGain")) {
+                agcGain = config->conf[name][getName()]["agcGain"];
+            }
             if (config->conf[name][getName()].contains("agcAttack")) {
                 agcAttack = config->conf[name][getName()]["agcAttack"];
             }
@@ -31,6 +37,8 @@ namespace demod {
 
             // Define structure
             demod.init(input, dsp::demod::SSB<dsp::stereo_t>::Mode::LSB, bandwidth, getIFSampleRate(), agcAttack / getIFSampleRate(), agcDecay / getIFSampleRate());
+            demod.setAGCEnabled(agcEnabled);
+            demod.setAGCGain(powf(10.0f, agcGain / 20.0f));
         }
 
         void start() { demod.start(); }
@@ -39,6 +47,30 @@ namespace demod {
 
         void showMenu() {
             float menuWidth = ImGui::GetContentRegionAvail().x;
+            if (ImGui::Checkbox(("AGC##_radio_lsb_agc_ena_" + name).c_str(), &agcEnabled)) {
+                demod.setAGCEnabled(agcEnabled);
+                _config->acquire();
+                _config->conf[name][getName()]["agcEnabled"] = agcEnabled;
+                if (!agcEnabled) {
+                    // Keep the last AGC gain as the manual gain
+                    _config->conf[name][getName()]["agcGain"] = agcGain;
+                }
+                _config->release(true);
+            }
+            if (agcEnabled) {
+                agcGain = std::clamp<float>(20.0f * log10f(demod.getAGCGain()), -10.0f, 90.0f);
+                ImGui::BeginDisabled();
+            }
+            ImGui::SameLine();
+            ImGui::SetNextItemWidth(menuWidth - ImGui::GetCursorPosX());
+            if (ImGui::SliderFloat(("##_radio_lsb_gain_" + name).c_str(), &agcGain, -10.0f, 90.0f, "%.0f dB")) {
+                demod.setAGCGain(powf(10.0f, agcGain / 20.0f));
+                _config->acquire();
+                _config->conf[name][getName()]["agcGain"] = agcGain;
+                _config->release(true);
+            }
+            if (agcEnabled) { ImGui::EndDisabled(); }
+            else { ImGui::BeginDisabled(); }
             ImGui::LeftLabel("AGC Attack");
             ImGui::SetNextItemWidth(menuWidth - ImGui::GetCursorPosX());
             if (ImGui::SliderFloat(("##_radio_lsb_agc_attack_" + name).c_str(), &agcAttack, 1.0f, 200.0f)) {
@@ -55,6 +87,7 @@ namespace demod {
                 _config->conf[name][getName()]["agcDecay"] = agcDecay;
                 _config->release(true);
             }
+            if (!agcEnabled) { ImGui::EndDisabled(); }
         }
 
         void setBandwidth(double bandwidth) { demod.setBandwidth(bandwidth); }
@@ -86,6 +119,8 @@ namespace demod {
 
         ConfigManager* _config;
 
+        bool agcEnabled = true;
+        float agcGain = 0.0f;
         float agcAttack = 50.0f;
         float agcDecay = 5.0f;